requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Pillow==10.0.0
opencv-python==4.8.0.76
wand==0.6.11
cairosvg==2.7.0
pdf2image==1.16.3
rawpy==0.18.1
svglib==1.5.1
reportlab==4.0.4
python-magic==0.4.27
gunicorn==21.2.0
pytest==7.4.0
python-dotenv==1.0.0
flask-cors==4.0.0 